surfaceflinger: change usageBits type to uint64_t
with uint32_t, it will loose the bits and making the usageBits invalid.

+++ services/surfaceflinger/CompositionEngine/include/compositionengine/FodExtension.h
@@ -23,6 +23,6 @@
 #define FOD_TOUCHED_LAYER_NAME "Fingerprint on display.touched#0"
 
 extern uint32_t getFodZOrder(uint32_t z, bool touched);
-extern uint32_t getFodUsageBits(uint32_t usageBits, bool touched);
+extern uint64_t getFodUsageBits(uint64_t usageBits, bool touched);
 
 #endif /* __FOD_EXTENSION__H__ */
